refactor: 외부 시스템(Google, Apple, Redis 등)에 대한 구체 구현체를 infrastructure 계층으로 이동
