+++ src/main/java/com/prography/lighton/auth/infrastructure/oauth/apple/AppleKeyUtils.java
@@ -1,4 +1,4 @@
-package com.prography.lighton.auth.infrastructure;
+package com.prography.lighton.auth.infrastructure.oauth.apple;
 
 import com.prography.lighton.auth.application.exception.ApplePrivateKeyLoadException;
 import com.prography.lighton.common.exception.base.NotFoundException;
