+++ src/main/java/com/prography/lighton/region/infrastructure/repository/RegionRepository.java
@@ -1,4 +1,4 @@
-package com.prography.lighton.region.domain.repository;
+package com.prography.lighton.region.infrastructure.repository;
 
 import com.prography.lighton.region.domain.entity.Region;
 import org.springframework.data.jpa.repository.JpaRepository;
